# src/main/resources/static/api-docs.yaml
openapi: 3.1.0
info:
  title: SocialMedia OpenApi Definition
  description: Open Api for SocialMedia project
  version: '1'
servers:
  - url: http://127.0.0.1:8081/api
    description: Generated server url
...
paths:
...
  /friends/invite:
    post:
      tags:
        - Friends
      summary: Invite user to friends
      operationId: invite
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: invited!
        '400':
          description: BAD_REQUEST
          content:
            text/plain:
              schema:
                type: string
                example: This is you!
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: User not found!
      security:
        - bearerAuth: []
  /chats/{id}:
    post:
      tags:
        - Chat & Message API
      summary: Create a new chat with user (id)
      operationId: createChat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: CREATED
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Chat'
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: User not found!
      security:
        - bearerAuth: []
    delete:
      tags:
        - Chat & Message API
      summary: Delete chat (id)
      operationId: deleteChat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
        '403':
          description: FORBIDDEN
          content:
            text/plain:
              schema:
                type: string
                example: Access denied!
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: Chat not found!
      security:
        - bearerAuth: []
...
  /images/profile:
    patch:
      tags:
        - Image API
      summary: Set profile picture
      operationId: setProfilePicture
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                profilePicture:
                  type: string
                  format: binary
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: success
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: Image not found
        '500':
          description: INTERNAL_SERVER_ERROR
          content:
            text/plain:
              schema:
                type: object
      security:
        - bearerAuth: []
  /friends/deny:
    patch:
      tags:
        - Friends
      summary: Deny invite
      description: Deny a user's invitation (yours or the inviter's)
      operationId: deny
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: denied!
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: Invite not found!
      security:
        - bearerAuth: []
  /friends/accept:
    patch:
      tags:
        - Friends
      summary: Accept user to friends
      operationId: accept
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: accepted!
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: Invite not found!
      security:
        - bearerAuth: []
  /chats/read:
    patch:
      tags:
        - Chat & Message API
      summary: Read messages with ids(body)
      description: Change message status "read" to true
      operationId: readMessage
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int64
        required: true
      responses:
        '200':
          description: OK
        '403':
          description: FORBIDDEN
          content:
            text/plain:
              schema:
                type: string
                example: Access denied!
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: Message not found!
      security:
        - bearerAuth: []
...
  /images/{filename}:
    get:
      tags:
        - Image API
      summary: Get image by filename
      operationId: getImage
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: NOT FOUND
          content:
            text/plain:
              schema:
                type: string
                example: Image not found!
  /friends:
    get:
      tags:
        - Friends
      summary: Get all friends of an Authenticated user
      operationId: getAll
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
                uniqueItems: true
      security:
        - bearerAuth: []
    delete:
      tags:
        - Friends
      summary: Remove user from friends
      operationId: remove
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: removed!
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: TestFirstname is not your friend!
      security:
        - bearerAuth: []
  /friends/invites:
    get:
      tags:
        - Friends
      summary: Get users I've invited as friends
      operationId: getInvites
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
                uniqueItems: true
      security:
        - bearerAuth: []
  /friends/invited:
    get:
      tags:
        - Friends
      summary: Get users who invited me as friends
      operationId: getInvited
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
                uniqueItems: true
      security:
        - bearerAuth: []
  /chats:
    get:
      tags:
        - Chat & Message API
      summary: Get all chats of Authenticated user
      operationId: getChats
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
                  format: uuid
                uniqueItems: true
      security:
        - bearerAuth: []
  /chats/user:
    get:
      tags:
        - Chat & Message API
      summary: Get companion
      description: Return other chat member
      operationId: getUserByChat
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
        '403':
          description: FORBIDDEN
          content:
            text/plain:
              schema:
                type: string
                example: Access denied!
      security:
        - bearerAuth: []
  /chats/messages/{id}/{page}:
    get:
      tags:
        - Chat & Message API
      summary: Get sorted messages from chat (id) with pagination
      description: Messages are sent via WebSocket
      operationId: getMessages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageResponseDTO'
        '403':
          description: FORBIDDEN
          content:
            text/plain:
              schema:
                type: string
                example: Access denied!
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: Chat not found!
      security:
        - bearerAuth: []
...
  /chats/messages/{id}:
    delete:
      tags:
        - Chat & Message API
      summary: Delete message (id) from chat
      operationId: deleteMessage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: Message not found!
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    Chat:
      type: object
      properties:
        id:
          type: string
          format: uuid
        dateOfLastSend:
          type: string
          format: date-time
...
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        dateOfBirth:
          type: string
        profilePicture:
          type: string
        country:
          type: string
        private:
          type: boolean
        valid:
          type: boolean
        blocked:
          type: boolean
    MessageResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        senderId:
          type: integer
          format: int64
        dateOfSend:
          type: string
          format: date-time
        text:
          type: string
        read:
          type: boolean
  securitySchemes:
    bearerAuth:
      type: http
      in: header
      scheme: bearer
      bearerFormat: JWT